pyproject.toml
[tool.poetry]
name = "battleship"
version = "0.1.0"
description = ""
authors = ["Nikolay Pushkin <adwedelf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
sqlalchemy = "^2.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
